eatgo-customer-api/src/main/java/com/example/eatgo/service/RegionService.java
package com.example.eatgo.service;

import com.example.eatgo.domain.Region;
import com.example.eatgo.domain.RegionRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class RegionService {

    private final RegionRepository regionRepository;

    public List<Region> getRegions() {
        List<Region> regions = regionRepository.findAll();
        return regions;
    }

    @Transactional
    public Region addRegion(String name) {
        Region region = Region.builder()
                .name(name)
                .build();

        regionRepository.save(region);

        return region;
    }

}
